refactor getCardDataShape to own file
... preventing compile link to .vue files
