+++ packages/vue/src/db/courseDB.ts
@@ -1,4 +1,5 @@
 import { DataShape } from '../base-course/Interfaces/DataShape';
+import Courses from '../courses';
 import { NameSpacer, ShapeDescriptor } from '../courses/NameSpacer';
 import ENV from '../ENVIRONMENT_VARS';
 import { CourseConfig } from '../server/types';
@@ -432,38 +433,6 @@ export async function getCourseQuestionTypes(courseID: string) {
 export async function getCourseConfig(courseID: string) {
   const config = await getCourseConfigs([courseID]);
   return config.rows[0].doc;
-}
-
-export async function getCardDataShape(courseID: string, cardID: string) {
-  const dataShapes: DataShape[] = [];
-  Courses.courses.forEach((course) => {
-    course.questions.forEach((question: any) => {
-      question.dataShapes.forEach((ds: any) => {
-        dataShapes.push(ds);
-      });
-    });
-  });
-
-  // log(`Datashapes: ${JSON.stringify(dataShapes)}`);
-
-  const db = await getCourseDB(courseID);
-  const card = await db.get<CardData>(cardID);
-  const disp = await db.get<DisplayableData>(card.id_displayable_data[0]);
-  const cfg = await db.get<CourseConfig>('CourseConfig');
-
-  // log(`Config: ${JSON.stringify(cfg)}`);
-  // log(`DisplayableData: ${JSON.stringify(disp)}`);
-
-  const dataShape = cfg!.dataShapes.find((ds) => {
-    return ds.name === disp.id_datashape;
-  });
-
-  const ret = dataShapes.find((ds) => {
-    return ds.name === NameSpacer.getDataShapeDescriptor(dataShape!.name).dataShape;
-  })!;
-
-  log(`Returning ${JSON.stringify(ret)}`);
-  return ret;
 }
 
 function getTagID(tagName: string): string {
